Move everything into atom_bindings
