+++ filenames.gypi
@@ -44,7 +44,6 @@
       'lib/common/api/deprecations.js',
       'lib/common/api/exports/electron.js',
       'lib/common/api/native-image.js',
-      'lib/common/api/process-stats.js',
       'lib/common/api/shell.js',
       'lib/common/init.js',
       'lib/common/reset-search-paths.js',
@@ -293,7 +292,6 @@
       'atom/common/api/atom_api_native_image.cc',
       'atom/common/api/atom_api_native_image.h',
       'atom/common/api/atom_api_native_image_mac.mm',
-      'atom/common/api/atom_api_process_stats.cc',
       'atom/common/api/atom_api_shell.cc',
       'atom/common/api/atom_api_v8_util.cc',
       'atom/common/api/atom_bindings.cc',
