+++ lib/common/init.js
@@ -46,5 +46,3 @@ if (process.type === 'browser') {
 if (process.platform === 'win32' && __dirname.indexOf('\\Program Files\\WindowsApps\\') === 2) {
   process.windowsStore = true
 }
-
-Object.assign(Object.getPrototypeOf(process), process.atomBinding('process_stats'))
